service/ledger/requirements.txt
celery==5.2.7
django-health-check==3.17.0
psycopg2-binary==2.9.6
